pyproject.toml
[tool.poetry]
name = "clustercheck"
version = "0.1.0"
description = "check a cluster of services"
authors = ["erik aronesty <erik@q32.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.4.1"
requests = "^2.25.1"
websocket-client = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"